Minimally functional is_ladder_capture function implemented.
Still needs to be evaluated with unit tests, and needs to cache away ladders for future moves if players decide to initiate the detected ladder, to eliminate redundant computation. Must especially test robustness of logic that handles decision of where to play hunter's next move during roll-out. Currently just looks for a move that is diagonal from previous hunter move which is also in one of prey's two liberties.